Use BEM for CSS naming and split board in two files
